type_changed

+++ summ.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 int main(){
-    int a,b; // Type can be probably changed
+    uint64_t a,b; // Carefully with var < 0
     cin>>a>>b;
     cout<<a+b<<endl;
 }
